+++ CSharp/TfsCmdlets.Shared/Cmdlets/CmdletBase.cs
@@ -10,7 +10,7 @@ namespace TfsCmdlets.Cmdlets
     /// </summary>
     public abstract class CmdletBase : PSCmdlet
     {
-        private static readonly IEnumerable<string> _valueReturningVerbs = new[] { "Get", "Test", "Invoke", "Export", "Search" };
+        private static readonly IEnumerable<string> _valueReturningVerbs = new[] { "Get", "Test", "Invoke", "Export", "Search", "Undo" };
 
         [Import] protected IPowerShellService PowerShell { get; set; }
         [Import] protected IParameterManager Parameters { get; set; }